requirements.txt
# Venice Flat Ontology Simulation Requirements
# Core dependencies for agent-based flood simulation

# Data processing and analysis
pandas>=2.0.0
numpy>=1.24.0
geopandas>=0.13.0

# Visualization
matplotlib>=3.6.0
seaborn>=0.12.0
plotly>=5.15.0

# Geospatial processing  
shapely>=2.0.0
fiona>=1.9.0
pyproj>=3.5.0

# LLM integration
openai>=1.3.0
python-dotenv>=1.0.0

# Scientific computing
scipy>=1.10.0
scikit-learn>=1.3.0

# Logging and utilities
loguru>=0.7.0
tqdm>=4.65.0
pathlib2>=2.3.7

# Data validation
jsonschema>=4.17.0
pydantic>=2.0.0

# Optional: Jupyter for interactive analysis
jupyter>=1.0.0
ipykernel>=6.22.0
nbformat>=5.8.0

# Optional: Advanced visualization
folium>=0.14.0
bokeh>=3.1.0

# Development tools
pytest>=7.3.0
black>=23.3.0
flake8>=6.0.0
